+++ Dappi.Cli/Helpers/ZipHelper.cs
@@ -3,7 +3,7 @@ using System.IO;
 using ICSharpCode.SharpZipLib.Core;
 using ICSharpCode.SharpZipLib.Zip;
 
-namespace Dappi.Cli;
+namespace Dappi.Cli.Helpers;
 
 public static class ZipHelper
 {
