+++ templates/MyCompany.MyProject.WebApi/MyCompany.MyProject.WebApi.csproj
@@ -9,11 +9,6 @@
   </PropertyGroup>
  
   <ItemGroup>
-    <Folder Include="Controllers\" />
-    <Folder Include="Entities\" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\..\CCApi.Extensions.DependencyInjection\CCApi.Extensions.DependencyInjection.csproj" />
     <ProjectReference Include="..\..\CCApi.SourceGenerator\CCApi.SourceGenerator.csproj" OutputItemType="Analyzer" />
   </ItemGroup>
